pyproject.toml
[tool.poetry]
name = "nelchan"
version = "0.1.0"
description = ""
authors = ["sh1ma <in9lude@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.2"
firebase-admin = "^5.0.0"
motor = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
pylint = "^2.8.3"
mypy = "^0.812"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
